Update the server resource to use the new style in resources
Adds an uninstall step like the tentacle resource

Includes a breaking change:
 - The remove step now deletes the instance and doesnt uninstall like it
previously did.
